Fix bug

+++ 2_finetune_init_lf.py
@@ -24,7 +24,7 @@ def use_embeddings_fasttext(model, word_embeddings):
     emb_tensor = torch.from_numpy(word_embeddings)
     size = emb_tensor.size()
     emb_class = nn.Embedding(size[0], size[1],padding_idx=0)
-    emb_class.weight = emb_tensor
+    emb_class.weight = emb_tensor.float()
     model.embeddings.word_embeddings = emb_class
 
     return model
